Move pytensor-cache to pytensor package
More or less a copy of https://github.com/aesara-devs/aesara/commit/96eca11f8822038eb4dc9978a212eeb24501763f
