+++ bin/pytensor-cache
@@ -1,3 +1,12 @@
 #!/usr/bin/env python
-import pytensor_cache
-pytensor_cache.main()
+import warnings
+
+from pytensor.bin.pytensor_cache import main
+
+
+warnings.warn(
+    message= "Using this bin/pytensor-cache script is deprecated. Use the plain "
+    "pytensor-cache command which is installed along with pytensor.",
+    category=DeprecationWarning,
+)
+main()
